// backend/src/types/ErrorCode.ts
// src/types/ErrorCodes.ts
export enum ErrorCode {
	REQUIRED_FIELDS_MISSING = 'REQUIRED_FIELDS_MISSING',
	EMAIL_EXISTS = 'EMAIL_EXISTS',
	INVALID_EMAIL = 'INVALID_EMAIL',
	INVALID_PASSWORD = 'INVALID_PASSWORD',
	PASSWORD_MISMATCH = 'PASSWORD_MISMATCH',
	USER_NOT_FOUND = 'USER_NOT_FOUND',
	INCORRECT_PASSWORD = 'INCORRECT_PASSWORD',
	ACCOUNT_NOT_ACTIVATED = 'ACCOUNT_NOT_ACTIVATED',
	OTP_EXPIRED = 'OTP_EXPIRED',
	OTP_INCORRECT = 'OTP_INCORRECT',
	OTP_VERIFICATION_FAILED = 'OTP_VERIFICATION_FAILED',
	ACCOUNT_ALREADY_ACTIVATED = 'ACCOUNT_ALREADY_ACTIVATED',
	UNVERIFIED_FOR_PASSWORD_RESET = 'UNVERIFIED_FOR_PASSWORD_RESET',
	RESET_PASSWORD_FAILED = 'RESET_PASSWORD_FAILED',
	OTP_NOT_RECEIVED = 'OTP_NOT_RECEIVED',
	UNKNOWN_ERROR = 'UNKNOWN_ERROR',
	SERVER_ERROR = 'SERVER_ERROR',
	DB_ERROR = 'DB_ERROR'
}